# chore: optimize config_detail view using materialized views