Previously:
```
 Planning Time: 2.682 ms
 Execution Time: 173.179 ms
 ```

After change:
```
 Planning Time: 2.133 ms
 Execution Time: 4.294 ms
```

Verified with multiple config items

Also, this query takes 2nd highest IO time in AWS cluster

![image](https://github.com/user-attachments/assets/230d1758-2c5f-49d7-bfa9-df422409e0b4)


Fixes: https://github.com/flanksource/duty/issues/1141